Move MenuItemPaused to components folder
